Try to use okhttp's default caching mechanisms
what the fuck
