+++ app/src/main/java/com/ldlda/chesscom_stats/HallOfFameFragment.java
@@ -18,6 +18,7 @@ import com.ldlda.chesscom_stats.adapter.HallOfFameAdapter;
 import com.ldlda.chesscom_stats.api.data.LeaderboardEntry;
 import com.ldlda.chesscom_stats.api.repository.ChessRepositoryJava;
 import com.ldlda.chesscom_stats.databinding.FragmentHallOfFameBinding;
+import com.ldlda.chesscom_stats.utils.UtilsKt;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -46,7 +47,10 @@ public class HallOfFameFragment extends Fragment {
             startActivity(intent);
         });
         recyclerView.setAdapter(adapter);
-        repo = new ChessRepositoryJava();
+        repo = UtilsKt.defaultWithCache(UtilsKt.buildCache(
+                requireContext().getCacheDir(), "http-cache",
+                20 << 20
+        ));
         swipeRefreshLayout.setOnRefreshListener(() -> {
             long now = System.currentTimeMillis();
             if (now - lastRefreshAt < MIN_REFRESH_INTERVAL_MS) {
