+++ app/src/main/java/com/ldlda/chesscom_stats/api/repository/ChessRepositoryImpl.kt
@@ -1,65 +1,22 @@
 package com.ldlda.chesscom_stats.api.repository
 
-import androidx.annotation.VisibleForTesting
 import com.ldlda.chesscom_stats.api.data.CountryInfo
 import com.ldlda.chesscom_stats.api.data.Leaderboards
 import com.ldlda.chesscom_stats.api.data.Player
 import com.ldlda.chesscom_stats.api.data.PlayerStats
 import com.ldlda.chesscom_stats.api.fetch.ChessApiClient
-import com.ldlda.chesscom_stats.api.fetch.DefaultChessApi
-import com.ldlda.chesscom_stats.utils.cache.TimedCache
+import java.net.URI
 
-class ChessRepositoryImpl(
-    private val client: ChessApiClient = DefaultChessApi,
-    private val playerCache: TimedCache<Player> = TimedCache(ttlMillis = 5 * 60_000L), // 5 min
-    private val statsCache: TimedCache<PlayerStats> = TimedCache(ttlMillis = 2 * 60_000L), // 2 min
-    private val leaderboardsCache: TimedCache<Leaderboards> = TimedCache(ttlMillis = 15 * 60_000L), // 15 min
-    private val countryCache: TimedCache<CountryInfo> = TimedCache(ttlMillis = 24 * 60 * 60_000L), // 24h
+// bare ass chess repo
+class ChessRepositoryImpl(val client: ChessApiClient) : ChessRepository {
+    override suspend fun getPlayer(username: String): Player =
+        client.getPlayer(username)
 
-) : ChessRepository {
-    private companion object {
-        private const val LEADERBOARDS_KEY = "leaderboards"
-    }
+    override suspend fun getPlayerStats(username: String): PlayerStats =
+        client.getPlayerStats(username)
 
-    override suspend fun getPlayer(username: String): Player {
-        val key = username.normalize()
-        playerCache.get(key)?.let { return it }
-        val player = client.getPlayer(key)
-        playerCache.put(key, player)
-        return player
-    }
+    override suspend fun getLeaderboards(): Leaderboards = client.getLeaderboards()
 
-    override suspend fun getPlayerStats(username: String): PlayerStats {
-        val key = username.normalize()
-        statsCache.get(key)?.let { return it }
-        val stats = client.getPlayerStats(key)
-        statsCache.put(key, stats)
-        return stats
-    }
-
-    override suspend fun getLeaderboards(): Leaderboards {
-        leaderboardsCache.get(LEADERBOARDS_KEY)?.let { return it }
-        val boards = client.getLeaderboards()
-        leaderboardsCache.put(LEADERBOARDS_KEY, boards)
-        return boards
-    }
-
-
-    override suspend fun getCountry(countryUrl: java.net.URI): CountryInfo {
-        val url = countryUrl.toString()
-        countryCache.get(url)?.let { return it }
-        val info = client.getCountryByUrl(url)
-        countryCache.put(url, info)
-        return info
-    }
-
-    private fun String.normalize(): String = trim().lowercase()
-
-    @VisibleForTesting
-    fun clearCaches() {
-        playerCache.clear()
-        statsCache.clear()
-        leaderboardsCache.clear()
-    }
+    override suspend fun getCountry(countryUrl: URI): CountryInfo =
+        client.getCountryByUrl(countryUrl.toString())
 }
-
